Moved Change class inside of Buffer because it's only accessed internally.
