+++ docs/javadoc/com/meti/buffer/Change.html
@@ -104,7 +104,7 @@ var activeTableTab = "activeTableTab";
         <li>java.lang.Object</li>
         <li>
             <ul class="inheritance">
-                <li>com.meti.io.buffer.Change</li>
+                <li>com.meti.io.buffer.Buffer.Change</li>
             </ul>
         </li>
     </ul>
@@ -128,7 +128,8 @@ implements java.io.Serializable</pre>
                     <dt><span class="simpleTagLabel">Author:</span></dt>
                     <dd>SirMathhman</dd>
                     <dt><span class="seeLabel">See Also:</span></dt>
-                    <dd><a href="../../../serialized-form.html#com.meti.io.buffer.Change">Serialized Form</a></dd>
+                    <dd><a href="../../../serialized-form.html#com.meti.io.buffer.Buffer.Change">Serialized Form</a>
+                    </dd>
                 </dl>
             </li>
         </ul>
