+++ docs/javadoc/com/meti/buffer/class-use/Change.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
     <!-- Generated by javadoc (1.8.0_101) on Fri Jan 05 14:02:48 EST 2018 -->
-    <title>Uses of Class com.meti.io.buffer.Change</title>
+    <title>Uses of Class com.meti.io.buffer.Buffer.Change</title>
     <meta name="date" content="2018-01-05">
     <link rel="stylesheet" type="text/css" href="../../../../stylesheet.css" title="Style">
     <script type="text/javascript" src="../../../../script.js"></script>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class com.meti.io.buffer.Change";
+            parent.document.title="Uses of Class com.meti.io.buffer.Buffer.Change";
         }
     }
     catch(err) {
@@ -70,9 +70,10 @@
     </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-    <h2 title="Uses of Class com.meti.io.buffer.Change" class="title">Uses of Class<br>com.meti.io.buffer.Change</h2>
+    <h2 title="Uses of Class com.meti.io.buffer.Buffer.Change" class="title">Uses of Class<br>com.meti.io.buffer.Buffer.Change
+    </h2>
 </div>
-<div class="classUseContainer">No usage of com.meti.io.buffer.Change</div>
+<div class="classUseContainer">No usage of com.meti.io.buffer.Buffer.Change</div>
 <!-- ======= START OF BOTTOM NAVBAR ====== -->
 <div class="bottomNav"><a name="navbar.bottom">
     <!--   -->
