+++ docs/javadoc/serialized-form.html
@@ -79,10 +79,10 @@
         <li class="blockList">
             <h2 title="Package">Package&nbsp;com.meti.io.buffer</h2>
             <ul class="blockList">
-                <li class="blockList"><a name="com.meti.io.buffer.Change">
+                <li class="blockList"><a name="com.meti.io.buffer.Buffer.Change">
                     <!--   -->
                 </a>
-                    <h3>Class <a href="com/meti/buffer/Change.html" title="class in com.meti.io.buffer">com.meti.io.buffer.Change</a>
+                    <h3>Class <a href="com/meti/buffer/Change.html" title="class in com.meti.io.buffer">com.meti.io.buffer.Buffer.Change</a>
                         extends java.lang.Object implements Serializable</h3>
                     <ul class="blockList">
                         <li class="blockList">
